new class dor user interface is added

+++ ResidualMaterials/Balance.cs
@@ -4,9 +4,9 @@
     {
         public int BalanceID { get; set; }
         public bool Type { get; set; }
-        public int Dim { get; set; }
-        public int Length { get; set; }
-        public int W { get; set; }
-        public int H { get; set; }
+        public double Dim { get; set; }
+        public double Length { get; set; }
+        public double W { get; set; }
+        public double H { get; set; }
     }
 }
